+++ packages/core/client/src/schema-component/antd/page/index.ts
@@ -2,3 +2,4 @@ export * from './Page';
 export * from './FixedBlock';
 export * from './PageTab.Settings';
 export * from './Page.Settings';
+export * from './FixedBlockDesignerItem';
